# Use native JSON instead of MessagePack is faster