--- a/package.json
+++ b/package.json
@@ -12,7 +12,7 @@
   "dependencies": {
     "async": "0.9.0",
     "debug": "2.2.0",
-    "redis": "2.4.2",
+    "redis": "^2.5.3",
     "socket.io-adapter": "0.4.0",
     "uid2": "0.0.3"
   },