Test source:
```javascript
var msgpackLite = require('msgpack-lite');
var msgpackJs = require('msgpack-js');
var Benchmark = require('benchmark');
var msgpack = require('msgpack');
var fs = require('fs');
var _ = require('lodash');

var testData = {
    blablabla: {
        hello: 'World',
        country: 255
    },
    x: 0,
    y: 253745
};

var suite = new Benchmark.Suite;
suite.add('MessagePack encode/decode', function() {
        var encoded = msgpack.pack(testData);
        var decoded = msgpack.unpack(encoded);
    })
    .add('MessagePackLite encode/decode', function() {
        var encoded = msgpackLite.encode(testData);
        var decoded = msgpackLite.decode(encoded);
    })
    .add('MessagePackJS encode/decode', function() {
        var encoded = msgpackJs.encode(testData);
        var decoded = msgpackJs.decode(encoded);
    })
    .add('JSON encode/decode', function() {
        var encoded = JSON.stringify(testData);
        var decoded = JSON.parse(encoded);
    })
    .on('cycle', function(event) {
        console.log(String(event.target));
    })
    .on('complete', function() {
        console.log('Fastest is ' + this.filter('fastest').map('name'));
    })
    .run({ 'async': false });
```

Test results:
```
MessagePack encode/decode x 70,190 ops/sec ±2.38% (87 runs sampled)
MessagePackLite encode/decode x 75,994 ops/sec ±2.36% (86 runs sampled)
MessagePackJS encode/decode x 34,879 ops/sec ±2.74% (90 runs sampled)
JSON encode/decode x 371,367 ops/sec ±1.71% (91 runs sampled)
Fastest is JSON encode/decode
```