@didaxRedux 2016-05-06
The msgpack library is used to compress data, so fast encode/decode is balanced by bandwidth saves.