@didaxRedux 2016-05-06
Yes... you are right!
I have only a doubt about UInt64 support (see https://github.com/socketio/socket.io-redis/pull/105) but your analysis is clear.