@darrachequesne 2016-05-08
Hi! I think a PR had already been opened on this matter: https://github.com/socketio/socket.io-redis/pull/44. Unfortunately, as pointed out by @rauchg:

> JSON doesn't support binary data.

Note: maybe we should add a test for this :angel: